+++ pkg/components/confpulsar/v1/option_test.go
@@ -1,7 +1,8 @@
-package confpulsar
+package confpulsar_test
 
 import (
 	"testing"
 )
 
-func TestPulsarOption(t *testing.T) {}
+func TestPulsarOption(t *testing.T) {
+}